fix(classes): resolve django-stubs manager typing errors
Mypy (via django-stubs plugin) couldn't resolve the manager type for
ClassOffering.objects because ClassOfferingQuerySet was defined in a
separate managers module. That cascaded into "couldn't resolve related
manager 'classes'" errors on Category.classes and Instructor.classes.

Move the QuerySet inline with the model (matching the pattern used in
membership/billing) and parametrize it as QuerySet["ClassOffering"]
so django-stubs can infer the manager. classes/managers.py now re-exports
the QuerySet for backwards compatibility with any external imports.
